+++ sdk/keyvault/azure-keyvault-keys/azure/keyvault/keys/_client.py
@@ -131,7 +131,7 @@ class KeyClient(KeyVaultClientBase):
         policy = kwargs.pop("release_policy", None)
         if policy is not None:
             policy = self._models.KeyReleasePolicy(
-                encoded_policy=policy.encoded_policy, content_type=policy.content_type
+                encoded_policy=policy.encoded_policy, content_type=policy.content_type, immutable=policy.immutable
             )
         parameters = self._models.KeyCreateParameters(
             kty=key_type,
@@ -562,7 +562,7 @@ class KeyClient(KeyVaultClientBase):
         policy = kwargs.pop("release_policy", None)
         if policy is not None:
             policy = self._models.KeyReleasePolicy(
-                content_type=policy.content_type, encoded_policy=policy.encoded_policy
+                content_type=policy.content_type, encoded_policy=policy.encoded_policy, immutable=policy.immutable
             )
         parameters = self._models.KeyUpdateParameters(
             key_ops=kwargs.pop("key_operations", None),
@@ -681,7 +681,7 @@ class KeyClient(KeyVaultClientBase):
         policy = kwargs.pop("release_policy", None)
         if policy is not None:
             policy = self._models.KeyReleasePolicy(
-                content_type=policy.content_type, encoded_policy=policy.encoded_policy
+                content_type=policy.content_type, encoded_policy=policy.encoded_policy, immutable=policy.immutable
             )
         parameters = self._models.KeyImportParameters(
             key=key._to_generated_model(),